\name{IntroToBiocAnno-package}
\alias{IntroToBiocAnno-package}
\alias{IntroToBiocAnno}
\docType{package}
\title{
\packageTitle{IntroToBiocAnno}
}
\description{
\packageDescription{IntroToBiocAnno}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{IntroToBiocAnno}
\packageIndices{IntroToBiocAnno}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{IntroToBiocAnno}

Maintainer: \packageMaintainer{IntroToBiocAnno}
}
\references{
~~ Literature or other references for background information ~~
}

\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
